feat: modules solved
